Caching the Inverse of a Matrix
Matrix inversion is usually a costly computation and there may be some
benefit to caching the inverse of a matrix rather than compute it
repeatedly.
Here are a pair of functions that are used to create a special object
that
stores a matrix and caches its inverse.
This function creates a special "matrix" object that can cache its
inverse.
And the second function computes the inverse of the special "matrix"
created by
makeCacheMatrix at first. If the inverse has already been calculated and
the
matrix has not changed, then it will retrieve the inverse from the
cache.